feat: refactor DisplayPreviewImage component and update import path
