+++ app/components/admin/ListGalleryFiles.vue
@@ -3,7 +3,7 @@ import type { File } from '@@/shared/types/File'
 import { DataTable } from '#components'
 import DeleteFile from '@/components/admin/DeleteFile.vue'
 import UploadGalleryFile from '@/components/admin/UploadGalleryFile.vue'
-import DisplayPreviewImage from '@/components/user/DisplayPreviewImage.vue'
+import DisplayPreviewImage from '@/components/gallery/DisplayPreviewImage.vue'
 import createReadableDate from '@/composables/dateHelper'
 import useFileServerApi from '@/composables/useFileServerApi'
 import { useFilesQuery } from '@/queries/useFilesQuery'
