+++ app/components/gallery/DisplayPreviewImage.vue
